House Price - Kaggle Competition
The goal of this project was to predict housing prices.
What made this project interesting was the sheer amount of features that this dataset came with. This created a challenge of determining which features were most important, and which features could be manipulated to enhance the predictive power of the model used. In this case, the model used was a Random Forest.